feat: Support additionalResources on UWP
